+++ demo/src/main/java/com/example/demo/dao/UserDao.java
@@ -92,7 +92,7 @@ public class UserDao {
     }
 
     // 유저 정보 조회
-    public Map<String, Object> loginUser (int id) {
+    public Map<String, Object> loginUser (long id) {
         String sql = "SELECT id, email, nickname FROM user WHERE id = ?";
         return jdbcTemplate.queryForMap(sql, id);
     }
